even control of different use cases
