+++ src/main/java/no/hauglum/flightlog/service/Scraper.java
@@ -12,6 +12,7 @@ import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -27,32 +28,22 @@ import static no.hauglum.flightlog.domain.TakeOff.HOVEN_LOEN;
 
 @Service
 public class Scraper {
+
+    @Autowired
+    private DocumentFactory mDocumentFactory;
+
     public static final int INDEX_OF_TD_WITH_PILOT_INFO = 2;
     private Logger mLogger = LoggerFactory.getLogger(this.getClass().getName());
     public static final String USER_ID = "user_id";
 
-    public Document scrape(String url){
-        try {
-            return Jsoup.connect(url).get();
-        } catch (UnknownHostException e) {
-            throw new FatalException("Kan ikke koble til " + url, e);
-        }catch (HttpStatusException e){
-            throw new FatalException("Kan ikke koble til "+url+" pga feil kode " + e.getStatusCode(),e);
-        } catch (IOException e) {
-            throw new FatalException("Noe har gått galt " + url, e);
-        }
-    }
-
-
     public void scrapeFlightlog() {
-
         int startYear = 2016;
         scrapeTakeOff(startYear, HOVEN_LOEN);
     }
 
     private void scrapeTakeOff(int startYear, String takeOffId) {
         mLogger.info("Rapport for startsted med id " + takeOffId);
-        List<DocumentWrapper> documents = getLogForTakeOff(startYear, takeOffId);
+        List<DocumentWrapper> documents = mDocumentFactory.getLogForTakeOff(startYear, takeOffId);
         readDocuments(documents);
     }
 
@@ -61,9 +52,8 @@ public class Scraper {
         List<FlightDay> days = new ArrayList<>();
         HashMap<String, DayPass> dayPasses = new HashMap<String, DayPass>();
 
-
         for (DocumentWrapper dw : documents) {
-            Elements rows = getRowsInTable(dw.getDocument());
+            Elements rows = mDocumentFactory.getRowsInTable(dw.getDocument());
 
             FlightDay flightDay = null;
             for (Element row : rows) {
@@ -93,31 +83,6 @@ public class Scraper {
         mLogger.info("Antall unike piloter: " + pilots.size());
         mLogger.info("Antall dagspass " + dayPasses.size());
         mLogger.info("Rapport slutt");
-    }
-
-    private List<DocumentWrapper> getLogForTakeOff(int startYear, String takaOffId) {
-        List<DocumentWrapper> documents = new ArrayList<>();
-        for (int year = startYear; year < LocalDate.now().getYear() +1; year++) {
-
-            int offset = 0;
-            int pageSize = 1000;
-            for (int page = 0; page < 10; page++) {
-                offset = page * pageSize;
-                String url1 = "https://no.flightlog.org/fl.html?l=2&country_id=160&start_id=" + takaOffId + "&a=42&year=" + year + "&offset=" + offset;
-                DocumentWrapper dw = new DocumentWrapper(scrape(url1), year);
-                documents.add(dw);
-                if(getRowsInTable(dw.getDocument()).size() < 3)
-                    break;
-            }
-        }
-        return documents;
-    }
-
-    private Elements getRowsInTable(Document doc) {
-        Elements tables = doc.select("table");
-        int noOfTables = tables.size();
-        Element table = tables.get(noOfTables - 1);
-        return table.select("tr");
     }
 
     protected boolean isADayRow(Element element) {
